corrected DBContext build
