+++ curve-api/curve-api.Database/Models/CategoryComment.cs
@@ -9,7 +9,7 @@ namespace curve_api.Database.Models
         //Table
         public int Id { get; set; }
         public int UserId { get; set; }
-        public int ReviewId { get; set; }
+        public int CategoryId { get; set; }
         public string Subject { get; set; }
         public string Content { get; set; }
         //Nav
