+++ curve-api/curve-api.Database/Models/Review.cs
@@ -8,8 +8,8 @@ namespace curve_api.Database.Models
         //Table
         public int Id { get; set; }
         public DateTime ReviewDate { get; set; }
-        public int IdividualId { get; set; }
-        public int OverAllScore { get; set; }
+        public int IndividualId { get; set; }
+        public int OverallScore { get; set; }
         //Nav
         public Individual Individuals { get; set; }
         public List<Category> Categories { get; set; }
