+++ curve-api/curve-api.Database/Models/SubCategoryComment.cs
@@ -4,12 +4,12 @@ using System.Text;
 
 namespace curve_api.Database.Models
 {
-    public class SubCategoryComments
+    public class SubCategoryComment
     {
         //Table
         public int Id { get; set; }
         public int UserId { get; set; }
-        public int ReviewId { get; set; }
+        public int SubCategoryId { get; set; }
         public string Subject { get; set; }
         public string Content { get; set; }
         //Nav
